% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crqa_dyads.R
\name{crqa_dyads}
\alias{crqa_dyads}
\title{Vectorized rqa over a set of timeseries grouped by gid
single or multidimensional
figures out all possible dyads in set
and runs rqa on them}
\usage{
crqa_dyads(gdata, opt, ...)
}
\description{
Vectorized rqa over a set of timeseries grouped by gid
single or multidimensional
figures out all possible dyads in set
and runs rqa on them
}
