% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crqa_full.R
\name{crqa_full}
\alias{crqa_full}
\title{Runs through a set of standardized rqa metrics, by running crqa_trials() for each}
\usage{
crqa_full(
  data,
  analyses = c("HR", "EA", "HR.EA", "HRwin", "EAwin"),
  condcols = c("category", "item", "stim", "condition"),
  easample = 15,
  eawsize = 30,
  eawstep = 10,
  writecache = T,
  pinfo = NULL,
  ...
)
}
\arguments{
\item{data}{data list containing at least ed, labelled emotibit data after compiling and merging functions}

\item{analyses}{the set of CRQA analyses that will be run through}

\item{condcols}{column names that identify unique trials that will be analysed together with crqa. Null is only one trial per person}
}
\description{
Runs through a set of standardized rqa metrics, by running crqa_trials() for each
}
