% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crqa_means.R
\name{crqa_means}
\alias{crqa_means}
\title{Averages a table of pairwise comparisons over people (default), trials or conditions}
\usage{
crqa_means(
  data,
  average_over = "pid1",
  rqa_vars = c("cor", "RR", "DET", "NRLINE", "maxL", "L", "ENTR", "rENTR", "LAM", "TT")
)
}
\description{
Doubles the table first, ie looks at full square rather than upper triangle, then gets means of rqa_vars for average_over
}
