% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crqa_trials.R
\name{crqa_trials}
\alias{crqa_trials}
\title{Runs crqa, mdcrqa or mdrqa on pairs or groups}
\usage{
crqa_trials(
  data,
  condcols = c("category", "condition", "item", "stim"),
  pinfo = NULL,
  pid = "pid",
  edvs = NULL,
  opt = F,
  dyads = T,
  pseudo_dyads = F,
  announce = F,
  samplerate = 1,
  ...
)
}
\arguments{
\item{data}{ed or a data list containing ed, labelled emotibit data after compiling and merging functions}

\item{condcols}{column names that identify unique trials that will be analysed together with crqa. Null is only one trial per person}

\item{pinfo}{column for a person designator, for in out group differences}

\item{pid}{column name for unique subject ID}

\item{edvs}{list of DVs to crq process, can be multidimensional. if null does all}

\item{dyads}{if cross recurrence between dyads T, or F for multidimensional analysis}
}
\description{
Will group together items using condcols, or people using pinfo
}
