% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emotibit_compile.R
\name{emotibit_compile}
\alias{emotibit_compile}
\title{Runs through data folder, collating named emotibit DVs (edv) into one file}
\usage{
emotibit_compile(
  datafolder = NULL,
  edv = c("EA", "HR", "ACC"),
  hour_adjust = 0,
  minstart = "2020-01-01 01:00:01 BST",
  fix_timestamp_from_file = T,
  logging = T,
  ...
)
}
\arguments{
\item{datafolder}{a folder of raw .csv data and .json files from multiple sessions and sensors, if blank will ask for one with system dialog}

\item{edv}{All emotibit time series data that you require. These are files listed in the session folders as emotibit_data_XX.csv, where XX is the dv name}

\item{hour_adjust}{a correction for time zones / day light savings.}

\item{minstart}{if we have data that is time stamped before this point, then we're going to assume that the timestamps are wrong}

\item{fix_timestamp_from_file}{in the case that minstart finds a problem, try and reconstruct start times from filenames}

\item{logging}{save emotibit_data_comb.csv to disk and output logs of errors, plots of sessions. If FALSE just returns compiled data}
}
\description{
before getting this to work, you need the process_bpm.py in a folder
and you need to create a virtual environment that has python 3
}
