% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emotibit_session_folders.R
\name{emotibit_getsessionfolders}
\alias{emotibit_getsessionfolders}
\title{given a datafolder filepath, or named directory in working directory, or nothing
it returns a list of emotibit session folders}
\usage{
emotibit_getsessionfolders(datafolder)
}
\description{
given a datafolder filepath, or named directory in working directory, or nothing
it returns a list of emotibit session folders
}
