% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmotibit_heartpy.R
\name{emotibit_heartpy}
\alias{emotibit_heartpy}
\title{Run heartpy through data to regenerate heart rate}
\usage{
emotibit_heartpy(
  datafolder = NULL,
  heartpy_reparse = F,
  winsize = 8,
  heartpy_location = "/Users/dcr/Dropbox/Rwork/python/hp"
)
}
\arguments{
\item{datafolder}{a folder of raw .csv data and .json files from multiple sessions and sensors, if blank will ask for one with system dialog}

\item{heartpy_reparse}{force a re-parsing of data}

\item{winsize}{the window for calculating heart rate}

\item{location}{of the emotibit parser file on your computer}
}
\description{
before getting this to work, you need the process_bpm.py in a folder
and you need to create a virtual environment that has python 3
in that folder, and install heartpy.
This is what I ran:
  ###  python3 -m venv /Users/dcr/Dropbox/Rwork/python/hp
  ###  source venv/bin/activate
  ###  pip install heartpy
}
