% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emotibit_parser.R
\name{emotibit_parser}
\alias{emotibit_parser}
\title{Run emotibit data through parser to make files for each sensor}
\usage{
emotibit_parser(
  datafolder = NULL,
  reparse = F,
 
    parser_location = "/Applications/EmotiBitSoftware-macOS/EmotiBitDataParser.app/Contents/MacOS/EmotiBitDataParser"
)
}
\arguments{
\item{datafolder}{a folder of raw .csv data and .json files from multiple sessions and sensors, if blank will ask for one with system dialog}

\item{reparse}{force a re-parsing of data}

\item{location}{of the emotibit parser file on your computure}
}
\description{
Run emotibit data through parser to make files for each sensor
}
