% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emotibit_process.R
\name{emotibit_process}
\alias{emotibit_process}
\title{Organize, process and compiles emotibit raw data files and .json files}
\usage{
emotibit_process(datafolder = NULL, ...)
}
\arguments{
\item{datafolder}{a folder of raw .csv data and .json files from multiple sessions and sensors, if blank will ask for one with system dialog}

\item{...}{Arguments for emotibit_session_folders, emotibit_parser, emotibit_heartpy, emotibit_compile}
}
\description{
It starts with a folder that contains all the .csv and .json data files from emotibit SD cards,
and ends with the data parsed out and organised into folders, and compiled into a single data.table
You can pass
}
