% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emotibit_session_folders.R
\name{emotibit_session_folders}
\alias{emotibit_session_folders}
\title{Organize emotibit raw data files and .json files, into labelled session folders.
Will default to using the eyethink lab key to translate the emotibit serial IDs into soemthing more friendly}
\usage{
emotibit_session_folders(
  datafolder = NULL,
  sensor_id = "eyethink",
  destination = "emotibit_data"
)
}
\arguments{
\item{datafolder}{a folder of raw .csv data and .json files from multiple sessions and sensors, if blank will ask for one with system dialog}

\item{sensor_id}{key to use: "eyethink", "none" or name of file with headings eid and serialid. If blank or it can't find the file, it will ask for location}

\item{destination}{name of folder where sorted files end up. Set to NULL if overwriting}
}
\value{
the name of the (new) datafolder
}
\description{
Organize emotibit raw data files and .json files, into labelled session folders.
Will default to using the eyethink lab key to translate the emotibit serial IDs into soemthing more friendly
}
