% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyethinkphysiosync-package.R
\docType{package}
\name{eyethinkphysiosync-package}
\alias{eyethinkphysiosync}
\alias{eyethinkphysiosync-package}
\title{eyethinkphysiosync: Tools for processing physiological data and calculating synchrony}
\description{
Proceses data from emotibit sensors, labels with behavioural data from gorilla and runs cross recurrence analysis
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dcr-eyethink/eyethinkphysiosync}
  \item \url{https://dcr-eyethink.github.io/eyethinkphysiosync/}
  \item Report bugs at \url{https://github.com/dcr-eyethink/eyethinkphysiosync/issues}
}

}
\keyword{internal}
