% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gorilla_eid_corrections.R
\name{gorilla_eid_corrections}
\alias{gorilla_eid_corrections}
\title{corrects eid from list of pid throughout data packet}
\usage{
gorilla_eid_corrections(processed_gorilla, corrections)
}
\description{
corrects eid from list of pid throughout data packet
}
