% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gorilla_emotibit_eid.R
\name{gorilla_emotibit_eid}
\alias{gorilla_emotibit_eid}
\title{Get ID from gorilla questionnaire}
\usage{
gorilla_emotibit_eid(data = NULL, eidtask = "Emotibit_ID")
}
\arguments{
\item{data}{compiled gorilla data}
}
\value{
pid, eid, experiment start time and experiment version
}
\description{
Get ID from gorilla questionnaire
}
