% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gorilla_physio_presentation.R
\name{gorilla_physio_presentation}
\alias{gorilla_physio_presentation}
\title{Process gorilla data from task 'AV block multiple stim types'}
\usage{
gorilla_physio_presentation(
  data = NULL,
  datafolder = NULL,
  taskname = "AV block multiple stim types",
  condcols = c("stim", "condition", "category", "item"),
  qlist = "demographics",
  stim_screen = "stimuli",
  quantkey = NULL,
  post_task = NULL,
  post_task_link = "item",
  ...
)
}
\arguments{
\item{data}{a list of imported data from data_collator_gorilla()}

\item{datafolder}{if no data supplied, looks in the folder with a collection of gorilla downloads. If NULL it will ask for file location}

\item{condcols}{the names of columns that pick out a unique trial in the data}

\item{stim_screen}{the critical video stimuli are identified as rows where the Zone.Type is "content_video" and the  Screen.Name in gorilla is in this list}

\item{quantkey}{a data.table of Response (text that was on a likert button) and r (what number you want to assign)}
}
\value{
td (trial data) and pd (participant data)
}
\description{
Process gorilla data from task 'AV block multiple stim types'
}
