% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio_merge.R
\name{physio_merge}
\alias{physio_merge}
\title{Label physio data with trial info
labels the data stream with participant and condition info
based on eid and timestamp}
\usage{
physio_merge(
  processed_emotibit = NULL,
  processed_gorilla = NULL,
  td = NULL,
  sinfo = NULL,
  pd = NULL,
  sd = NULL,
  ed = NULL,
  condcols = c("category", "condition", "item", "stim"),
  normlevel = 1,
  pinfo = NULL,
  gd = NULL,
  baseline = FALSE,
  hour_adjust = 0,
  stop_at_plotcheck = F,
  eda_check = list(mean = 20, range_min = 0.01, low = 0.03, high = 25)
)
}
\arguments{
\item{processed_emotibit}{output from emotibit_process() or eomotibit_compile() if used}

\item{processed_gorilla}{output from gorilla_physio_presentation if used}

\item{td}{trial data, including pid, eid, cond(s), starttime and stoptime, if you don't have (processed_gorilla)}

\item{pd}{participant data, if you don't have (processed_gorilla)}

\item{sd}{stimuli data, if you don't have (processed_gorilla)}

\item{ed}{physio data, post compilation, if you don't have (processed_emotibit)}

\item{condcols}{the cols in td that pick out unique trial. If each person just saw the same single event, set to NULL}

\item{normlevel}{defaults to 1 which is zscoring per person. Set to 2+ for each additional column of condcols to use}

\item{pinfo}{name of a column in participant data that identifies a social group}

\item{baseline}{not implemented yet. do you want to z score using sd (F default) or just baseline shift (T)}

\item{hour_adjust}{nunmber of hours to shift physio data by, in case of GMT problem}

\item{eda_check}{set to NULL if not wanted, otherwise a list of mean, range_min, lod and high thresholds for EDA filtering}
}
\value{
labeled ed, td, pd and sd data with summary stats
}
\description{
Label physio data with trial info
labels the data stream with participant and condition info
based on eid and timestamp
}
