% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio_movie.R
\name{physio_movie}
\alias{physio_movie}
\title{Animates a physio_timeplot, scales and adds to stimulus video into a composite movie}
\usage{
physio_movie(
  timeplot = NULL,
  stimfile,
  outputfile = NULL,
  fps = 30,
  spwidth = 800,
  spheight = 400,
  ...
)
}
\arguments{
\item{timeplot}{a plot from function physio_timeplot, or pass variables for that function here and it will make a new one}

\item{stimfile}{the stimulus video - leave blank and you can select it with the file OS}

\item{outputfile}{name for composite movie. If blank, we'll just use the stim file name and add _spark}

\item{fps}{frame per second of movie}

\item{spwidth}{width of the sparkline movie - the sitmulus movie will be scaled to the same width before being composited}

\item{spheight}{height of the sparkline movie}

\item{...}{
  Arguments passed on to \code{\link[=physio_timeplot]{physio_timeplot}}
  \describe{
    \item{\code{data}}{data list from physio_merge}
    \item{\code{plotmetric}}{The metric in data$dm to plot, eg DET or norm}
    \item{\code{plotdv}}{The dv, eg HR or EA}
    \item{\code{selection}}{A datatable with the columns and values that you want to use to select the data to plot}
    \item{\code{smoother}}{parameter for loess smoothing. set to 0 for none}
    \item{\code{plot_error}}{give error from smoothing}
  }}
}
\description{
Animates a physio_timeplot, scales and adds to stimulus video into a composite movie
}
