% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio_merge.R
\name{physio_merge}
\alias{physio_merge}
\title{Label physio data with trial info
labels the data stream with participant and condition info
based on eid and timestamp}
\usage{
physio_merge(
  processed_emotibit = NULL,
  processed_gorilla = NULL,
  td = NULL,
  sinfo = NULL,
  pd = NULL,
  sd = NULL,
  ed = NULL,
  dinfo = c("pid"),
  normlevel = 1,
  pinfo = NULL,
  gd = NULL,
  filter_prop_min = 0.5,
  normvar = "filter",
  normvalue = 1,
  baseline = FALSE,
  hour_adjust = 0,
  trim_eda = T,
  stop_at_plotcheck = F,
  plot_eda_outliers = F
)
}
\arguments{
\item{td}{trial data, including eid, cond(s), starttime and stoptime}

\item{ed}{physio data, post compile}

\item{dinfo}{the cols in td that pick out unique trial. If single rec, just audience ID}

\item{normlevel}{how many of those dinfo cols should be used to norm}

\item{gd}{timecourse data that you also want labelled, optional wrapper for timecoursejoin_empatica}

\item{normvar}{use the period when normvar=normvalue for normalising}

\item{normvalue}{the raw value of normvar that you want to include in normalising}

\item{baseline}{do you want to z score using sd (F default) or just baseline shift (T)}

\item{hour_adjust}{nunmber of hours to shift physio data by, in case of GMT problem}
}
\value{
labeled ed data and td with summary stats
}
\description{
Label physio data with trial info
labels the data stream with participant and condition info
based on eid and timestamp
}
